Corrected the naming conventions
Corrected the naming conventions to camelCase. Also changed redundant model naming in most of the project
