--- ms-campus-kid-subjects/ExampleDocker/appsettings.json
@@ -1,13 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "FacultyConnection": "Server=host.docker.internal;Port=4007;Uid=root;Pwd=1234;Database=db_subject;"
-  }
-
-}
