+++ ms-campus-kid-subjects/SlnProjectMicroservice.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.3.32825.248
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "subjects-ms", "ExampleDocker\subjects-ms.csproj", "{A46AB254-2201-4B85-A80C-4E5CE7E4C038}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "subjects-ms", "subjects-ms\subjects-ms.csproj", "{A46AB254-2201-4B85-A80C-4E5CE7E4C038}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataBase", "DataBase\DataBase.csproj", "{62175055-7357-40E6-B93C-C77A9112BF6F}"
 EndProject
