+++ ms-campus-kid-subjects/subjects-ms/Program.cs
@@ -1,0 +1,45 @@
+using DataBase;
+using Microsoft.EntityFrameworkCore;
+using System.Data;
+
+var builder = WebApplication.CreateBuilder(args);
+
+// Add services to the container.
+
+builder.Services.AddControllers();
+// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+
+
+builder.Services.AddDbContext<SubjectsContext>(options => {
+    var conString = builder.Configuration.GetConnectionString("FacultyConnection");
+    options.UseMySql(conString, ServerVersion.AutoDetect(conString));
+}) ;
+
+    var app = builder.Build();
+
+
+
+using (var scope = app.Services.CreateScope())
+{
+    var context = scope.ServiceProvider.GetRequiredService<SubjectsContext>();
+    context.Database.Migrate();
+}
+
+
+
+// Configure the HTTP request pipeline.
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI();
+}
+
+app.UseHttpsRedirection();
+
+app.UseAuthorization();
+
+app.MapControllers();
+
+app.Run();
